[Paywalls V2] Extends support for blurred shadows to all Android versions (#1957)
